pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "qrcode-detect-yolo"
version = "0.1.0"
description = "Lightweight QR code detection pipeline scaffold for YOLO-based training and inference."
readme = "README.md"
requires-python = ">=3.11"
dependencies = []

[project.optional-dependencies]
inference = ["ultralytics>=8.3.0", "pillow>=10.4.0"]
synthetic = ["pillow>=10.4.0"]
dev = ["pytest>=8.3.0"]

[project.scripts]
qrcode-detect = "qrcode_detector.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
